+++ test/test_paradigm_snapshot.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # test_paradigm_snapshot.py
 # test DWDSmor paradigm snapshots for regression
-# Andreas Nolda 2024-08-22
+# Andreas Nolda 2024-09-02
 
 import io
 import csv
@@ -291,21 +291,21 @@ VERB_POS = "V"
 
 VERB_LEMMAS = ["spielen",    # VWeak
                "segeln",     # VWeak-el-er
-               "senden",     # VWeak, VInf, VPPres, VPPast-senden, VPres, VPastInd-haben, VPastSubjWeak, VImp
+               "senden",     # VWeak, VInf, VPPres, VPPast-d_t, VPres, VPastInd-d-t_t, VPastSubjWeak, VImp
                "denken",     # VInf, VPPres, VPPastWeak, VPres, VPastIndWeak, VPastSubjWeak, VImp
-               "haben",      # VInf, VPPres, VPPastWeak, VPresInd23Sg, VPresNonInd23Sg, VPastInd-haben, VPastSubj-haben, VImp
+               "haben",      # VInf, VPPres, VPPastWeak, VPresInd23Sg, VPresNonInd23Sg, VPastInd-d-t_t, VPastSubj-haben, VImp
                "wissen",     # VInf, VPPres, VPPastWeak, VModPresIndSg, VModPresNonIndSg, VPastIndWeak, VPastSubjWeak
                "können",     # VInf, VPPres, VPPastWeak, VPPastStr, VModPresIndSg, VModPresNonIndSg, VPastIndWeak, VPastSubjWeak
                "gehen",      # VInf, VPPres, VPPastStr, VPres, VPastStr, VImp
                "schwimmen",  # VInf, VPPres, VPPastStr, VPres, VPastIndStr, VPastSubjStr, VPastSubjOld, VImp
-               "laden",      # VInf, VPPres, VPPastStr, VPres, VPresInd23Sg-laden, VPresNonInd23Sg, VPastIndStr, VPastSubjStr, VImp
+               "laden",      # VInf, VPPres, VPPastStr, VPres, VPresInd23Sg-d_t, VPresNonInd23Sg, VPastIndStr, VPastSubjStr, VImp
                "laufen",     # VInf, VPPres, VPPastStr, VPresInd23Sg, VPresNonInd23Sg, VPastStr, VImp
-               "halten",     # VInf, VPPres, VPPastStr, VPresInd23Sg-t, VPresNonInd23Sg, VPastStr, VImp
+               "halten",     # VInf, VPPres, VPPastStr, VPresInd23Sg-t_0, VPresNonInd23Sg, VPastStr, VImp
                "sehen",      # VInf, VPPres, VPPastStr, VPresInd23Sg, VPresNonInd23Sg, VPastIndStr, VPastSubjStr, VImpSg, VImpPl
-               "tun",        # VInf-tun, VPPres, VPPast-tun, VPres-tun, VPastIndStr, VPastSubjStr, VImpSg0, VImpPl,
+               "tun",        # VInf_n, VPPres, VPPast_n, VPres-tun, VPastIndStr, VPastSubjStr, VImpSg0, VImpPl,
                "werden",     # VInf, VPPres, VPPastStr, VPresInd2Sg-werden, VPresInd3Sg-werden, VPresNonInd23Sg, VPastInd-werden,
                              # VPastIndSg-ward, VPastIndPl-werden, VPastSubjStr, VImp
-               "sein"]       # VInf-tun, VPPres, VPPastStr, VPresInd1Sg-sein, VPresInd2Sg-sein, VPresInd3Sg-sein, VPresInd13Pl-sein,
+               "sein"]       # VInf_n, VPPres, VPPastStr, VPresInd1Sg-sein, VPresInd2Sg-sein, VPresInd3Sg-sein, VPresInd13Pl-sein,
                              # VPresInd2Pl-sein, VPresSubj-sein, VPastIndStr, VPastSubjStr, VPastSubj2-sein, VImpSg0, VImpPl-sein
 
 @fixture